Oppdaterer IM builder
